+++ Reflection/src/utopia/reflection/component/reach/template/CursorDefining.scala
@@ -1,7 +1,8 @@
 package utopia.reflection.component.reach.template
 
-import utopia.genesis.shape.shape2D.Bounds
-import utopia.reflection.cursor.CursorType
+import utopia.genesis.image.Image
+import utopia.genesis.shape.shape2D.{Bounds, Point}
+import utopia.reflection.cursor.{Cursor, CursorType}
 
 /**
   * A common trait for components that specify what type of cursor should be used inside their bounds
@@ -13,11 +14,19 @@ trait CursorDefining
 	/**
 	  * @return The type of cursor displayed over this component
 	  */
-	def cursor: CursorType
+	def cursorType: CursorType
 	
 	/**
 	  * @return The area <b>relative to the cursor managed context</b> defined by this component. Usually this means
 	  *         the component's location and size in relation to the managed canvas element.
 	  */
 	def cursorBounds: Bounds
+	
+	/**
+	  * Determines the image to use for the cursor
+	  * @param cursor Proposed cursor data
+	  * @param position Position of the cursor, relative to this component's cursorBounds property
+	  * @return The image to draw as the cursor over that location
+	  */
+	def cursorToImage(cursor: Cursor, position: Point): Image
 }