+++ Reflection/src/utopia/reflection/cursor/ReachCursorManager.scala
@@ -1,6 +1,8 @@
 package utopia.reflection.cursor
 
-import utopia.genesis.shape.shape2D.Point
+import utopia.flow.util.CollectionExtensions._
+import utopia.genesis.shape.shape2D.{Bounds, Point}
+import utopia.reflection.color.ColorShadeVariant
 import utopia.reflection.component.reach.template.CursorDefining
 
 /**
@@ -8,7 +10,7 @@ import utopia.reflection.component.reach.template.CursorDefining
   * @author Mikko Hilpinen
   * @since 11.11.2020, v2
   */
-class ReachCursorManager(cursors: CursorSet)
+class ReachCursorManager(val cursors: CursorSet)
 {
 	// ATTRIBUTES	-----------------------------
 	
@@ -19,12 +21,28 @@ class ReachCursorManager(cursors: CursorSet)
 	
 	/**
 	  * @param position A position in the cursor managed context
+	  * @param shadeOf A function for calculating the overall shade of the targeted area
 	  * @return Cursor image to use over that position
 	  */
-	def cursorAt(position: Point) = cursorComponents.find { _.cursorBounds.contains(position) } match
+	def cursorAt(position: Point)(shadeOf: Bounds => ColorShadeVariant) =
 	{
-		case Some(component) => cursors(component.cursor)
-		case None => cursors.default
+		// Checks whether any of the registered components is controlling the specified position
+		cursorComponents.findMap { c =>
+			val bounds = c.cursorBounds
+			if (bounds.contains(position))
+				Some(c -> bounds)
+			else
+				None
+		} match
+		{
+			// Case: Component manages area => lets the component decide cursor styling
+			case Some((component, bounds)) =>
+				component.cursorToImage(cursors(component.cursorType), position - bounds.position)
+			// Case: Cursor is outside all registered component zones => uses default cursor with modified shade
+			case None =>
+				val cursor = cursors.default
+				cursor(shadeOf(cursor.defaultBounds.translated(position)).opposite)
+		}
 	}
 	
 	/**